+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### v0.3
 #### Added
 - `invite` command.
+- Error when giveaway result is too big is handled.
 #### Changed
 - Using `giveaway` command requires "manage guild" permission.
 - Increased `meme`'s cooldown from 4s to 6s.